Fix stub_test for Mips64
Force ENTRY macro to load GP. Instead .cpload (used for o32 abi,
not working for n64) use .cpsetup to set up GP and save previous
GP value.

Fix expectations in GetSet32Static.

+++ runtime/arch/stub_test.cc
@@ -1836,7 +1836,7 @@ static void GetSet32Static(ArtField* f, Thread* self, mirror::ArtMethod* referre
                                            self,
                                            referrer);
 
-    EXPECT_EQ(res, values[i]) << "Iteration " << i;
+    EXPECT_EQ(static_cast<uint32_t>(res), values[i]) << "Iteration " << i;
   }
 #else
   UNUSED(f, self, referrer, test);
